requirements.txt
streamlit>=1.36.0
requests>=2.31.0
beautifulsoup4>=4.12.3
PyYAML>=6.0.2
python-dateutil>=2.9.0.post0
pytz>=2024.1
dateparser>=1.2.0
feedparser>=6.0.10
